All right, fine, so it didn't compile.  That's what I get for committing quick fixes from where I can't test them.
At any rate, now the detail screen will properly display minutes and seconds if need be.  It will also behave slightly differently with old data: It will remember the last proper data it found and rely on that if it can't get most recent data from LocationManager or that data is too old.  This does mean that if the last thing remembered was too old, it'll still use that, but the only significant way that'll happen is if the user stays on the preferences screen for more than two minutes from the detail screen.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@97 7ebfa8b6-0479-11de-af5c-836599625e1f