Changed the constructor for FinalDestinationOverlay so it takes an Info object instead of a GeoPoint.  It can get the GeoPoint from there, plus this gives it access to the Graticule of choice.
Also, created FinalDestinationDisabledOverlay.  This does the same thing as FDO for now, but will later handle clicks to switch to a nearby graticule.

MainMap was also updated to suit these changes.  Note that this will need to be changed later just in case we're Geohashing in Greenland, which, on the 30W line, does crazy things to the nearby final destination part.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@172 7ebfa8b6-0479-11de-af5c-836599625e1f