First steps of converting the stock grabbing into its own activity.  This, when finished, will remove a LOT of logic from the main class, which is Good(tm), though it will still be a bit of work before it's functioning properly.
To that end, StockGrabber is a new activity.  When called with an Intent containing a Graticule and a Calendar, it does all the stock grabby stuff via HashBuilder and returns a stock value (or an error) as a result.  This way, the main class won't need all the special-casing for if the activity is paused/destroyed while the stock grabbing is in progress.

It SHOULD look like a normal dialog if something needs to be done, but apparently that doesn't quite look right if I use Theme.Dialog.  Thus, I still need to rewrite it so it looks right (whether "right" is "like a dialog" or "right for this").

Also note that the main class is broken for now.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@176 7ebfa8b6-0479-11de-af5c-836599625e1f