As it turns out, if the Geohash Droid app is terminated in any way (most likely via low memory errors when it isn't active), the static mStore value in HashBuilder is rendered null.  However, the only time it's initialized is during the main GeohashDroid class.  Meaning, if, for instance, the user is in MainMap (a somewhat memory-heavy Activity due to it being a MapActivity), navigates to another app (i.e. the View Wiki Page option, or an incoming phone call), and the OS thinks memory is low, it could terminate Geohash Droid, expecting to start it back up from a retained state when it comes back in.
Unfortunately, if MainMap or whatnot needs the database at that time (such as the Nearby Points feature), this will crash, as mStore will be null, seeing as how HashBuilder.initialize won't have been run.

This fix is a quick hack that ensures mStore is always defined at any point it's needed by encapsulating it in an accessor that can create it from scratch if need be.  This does mean all calls that refer to mStore now need a Context, which has also been addressed here.

Later, a REAL fix would be to wrap all database access into a Service that knows to initialize HashBuilder without the hackish Context requirement every time.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@272 7ebfa8b6-0479-11de-af5c-836599625e1f