Resourceified the strings in WikiPictureEditor.
Also made it so that the first thing PictureEditor does is grab the URI to the image so that the cursor isn't as important to keep (it'll get destroyed during a config event).  There might be a better way to do this.

Next, I have to figure out why it seems to be picking the same picture every time no matter what I select from my phone.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@283 7ebfa8b6-0479-11de-af5c-836599625e1f