Initial commit of the new small screen support.  The small screen now has its own layout sections (layout-small-land and layout-small-port).  Right now, the only modified layouts are for the main screen, replacing the rather a bit huge DatePicker with a DateButton that leads to a DatePickerDialog.
This also adds in DateButton, a button that displays a DatePickerDialog and displays the currently selected date as its text field.

And THAT brought with it a change to the main GeohashDroid class, which allows it to try to figure out whether it's on the small or not-small screens by means of checking for the existence of the DatePicker, and acting accordingly.

While this does work for now, DateButton isn't preserving its state, meaning it'll reset back to the current date on an orientation shift and the dialog won't stay popped up if it was up.  I need to fix that.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@319 7ebfa8b6-0479-11de-af5c-836599625e1f