Fixed the regex to match the commented gallery section.
Also tried to add a regex to match the Photos section header if the <gallery> tag can't be found so that it gets added to the right place.  Unfortunately, there must be some tricks to Java's regex system that I'm missing, because the regex I have isn't matching properly yet.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@349 7ebfa8b6-0479-11de-af5c-836599625e1f