Backported r375:
HashBuilder.StockRunner will now look through a list of servers (currently just
two) just in case one doesn't know what the stock value for the day is (this
happens sometimes).  As such, now it goes to the Crox server first for data, as
that tends to be a bit more reliable and has its own redundancies to prevent bad
data.

Also, fixed a potential problem with writing the stock to the cache if the user
aborted the connection.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/0.7.4-backports@404 7ebfa8b6-0479-11de-af5c-836599625e1f