Oh, now I get it.  If I send a null as the string used to announce the notification, it doesn't announce itself, it just places itself in the notification bar.  That looks a lot smoother.
Yes, I know that from the comments I put in the code, the casual-language strings in the program, and commit messages like this, it looked like the "And away we go..." string was intentional.  Rest assured, it was not, and I was just trying to find a way around it.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@418 7ebfa8b6-0479-11de-af5c-836599625e1f