Started work on LocationAwareActivity.
So GeohashService didn't work as planned so far.  Fair enough.  In its
place, LocationAwareActivity is going up.  It will make a common Activity
base class for anything that should start listening for location updates at
resume() time and stop at pause() time.  This will go into the detail
screen and the two wiki screens (not the main map, since that needs a
different base class for the map itself).

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@505 7ebfa8b6-0479-11de-af5c-836599625e1f