Created QueueService.
Next up is the idea Thomas Hirsch had for queueing up wiki posts to send
later if the connection is bad.  To facilitate this, I'm working on making
a generic QueueService.  This is so that the attempts can be made in the
background, report back if any fatal errors happen (bad password, etc), and
so the queue can be observed and manipulated after the fact.  Should be
great, if I can pull it off the way I want to.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@511 7ebfa8b6-0479-11de-af5c-836599625e1f