Keep track of the location throughout WikiPictureEditor's execution.
Set it up so that the location of the image (that which should be posted
if the user checks the appropriate checkbox) is stored for later use by
the rest of WikiPictureEditor's execution thread.  Why, you ask?  So I
can use it later to stamp an infobox on the picture before uploading it,
all in the next few updates.

That's right.  I went there.  Booyah.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@522 7ebfa8b6-0479-11de-af5c-836599625e1f