Major layout reworking for the wiki screens.
First, the landscape picture upload screens have been reorganized so that
everything fits.  The image was moved to the right side since we don't use
the wide flickable gallery.  Everything else was reorganized to fit and to
make sure the small screen doesn't cut off text in any case.

Second, the portrait wiki posting screens were rearranged so that the
anonymous warning text will always fit properly.

Third, a large-size picture upload screen was added.  I've got more room to
work with on a large screen, so the text there is more verbose to make the
layout look better.

Two problems still to solve: Portrait small screen picture uploading won't
fit the anonymous warning, and for some reason it appears the G1 thinks it
is a large screen, meaning it displays the longer text and thus ruins the
layout in its limited space.  Must research.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@531 7ebfa8b6-0479-11de-af5c-836599625e1f