Fixed layout issues on small screens.
There.  That should do it.  Added a small hack in to the thumbnail size to
make the wiki picture uploader display right on small screens.

In the near future, I may want to also put some of the small-size dimens in
with sizedimens.  That may result in the size-specific layouts being
removed, too.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@533 7ebfa8b6-0479-11de-af5c-836599625e1f