Fleshed out QueueService even more.
Now QueueService should be ready for some testing.  Note that this implies
it hasn't been tested yet.  This implication is very very true, and is
being dealt with next.

If it turns out it DOES work through some as-yet-not-understood miracle of
computer science, I'll be making a Service to handle all wiki communication
so posts will queue up and run in the background, NOT block the UI until
each one is done, AND pause and automatically resume if there's no data
connection.  How 'bout THAT, huh?

So long story short, Thomas, your suggestion is finally being implemented.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@549 7ebfa8b6-0479-11de-af5c-836599625e1f