Redid QueueService so the thread doesn't wait on pause.
This version of QueueService is a bit more robust and sane.  It won't wait
the thread at pause time; rather, the thread will die and be restarted when
asked to resume.

Also, because there doesn't appear to be any sort of onSaveInstanceState
for Services, there are also methods to serialize Intents to disk.  These
need more flesh on them.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@550 7ebfa8b6-0479-11de-af5c-836599625e1f