Added a bunch of constants in to WikiPostService.
WikiPostService now has the constants for the extras that'll be used
when sending Intent data to it.  I think this covers all the options
that can currently be sent to it.  Next, I'll work on converting the
wiki posting commands to stuff that can happen in service of the
service.

THEN I go for the tricky business of figuring out error handling in the
service...

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@565 7ebfa8b6-0479-11de-af5c-836599625e1f