Implemented WikiMessageHandler.
I'm working on moving the logic for posting to the wiki out of the Activity
classes and into handlers which will be invoked by WikiPostService.  That
will allow for the entire "post in the background" bit to work.

To be honest, besides intent validation and stripping all the dialog stuff
out of the runners, this shouldn't be too hard.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@568 7ebfa8b6-0479-11de-af5c-836599625e1f