Fleshed out WikiMessageHandler even more.
First, WikiPostHandler is now WikiServiceHandler, it's abstract, and it now
exists outside of WikiPostService.

Second, WikiMessageHandler should, in theory, now be able to handle a post
from beginning to end.  However, in doing so, I realized that the exception
system coming back from WikiUtils doesn't give me enough information to
relay appropriate data back to WikiPostService.  That is, I can't tell from
the exceptions alone whether or not the service should wait for a data
connection or just pause entirely.  I'll need to rethink this.  Most likely
I'll just determine the problem in WikiUtils and make WikiExceptions out of
whatever it throws out.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@574 7ebfa8b6-0479-11de-af5c-836599625e1f