Made WikiServiceHandler.handlePost throw exceptions.
Instead of returning the status that WikiUtils gives it, WikiServiceHandler
now throws exceptions if anything goes wrong (that is, if it returns, the
service can assume all was well and it should continue).  WikiUtils hasn't
been updated to account for this yet, though, so it still needs to catch
Exception to allow it to compile properly.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@576 7ebfa8b6-0479-11de-af5c-836599625e1f