Converted all exceptions in WikiUtils to WikiException.
First off, the subclass idea for WikiException just to determine severity?
Yeah, that wasn't a good idea.  When that sort of information is vital,
it's best to force a check on it.  Otherwise, I'd either need to declare
all the subclasses as something WikiUtils can throw or I'd need to catch
the superclass, which just strikes me as sloppy and prone to future
problems.

To that end, all of WikiUtils's methods now throw WikiException.  And the
handler throws WikiException (and thus doesn't catch Exception anymore).
Much cleaner.  Once the service handler is set up to handle that, it'll be
great.

Plus, some formatting was done.  Whoops.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@578 7ebfa8b6-0479-11de-af5c-836599625e1f