Added a reset method to WikiBaseActivity.
WikiBaseActivity now has reset().  This resets the screen after a post.
This DOES seem an awful lot like doDismiss() right now, but with the switch
to WikiPostService for posts coming up, doDismiss will be going away, as
well as the entire dialog dismissing harness thingamajig that was calling
it in the first place.  So the duplicate code in WikiPictureEditor will go
away, too.

Oh, and reset also wipes out the text.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@590 7ebfa8b6-0479-11de-af5c-836599625e1f