Made sure the Submit Message button is disabled if there's no text.
There's really no reason why the Submit Message button should be available
if there's no text in the message poster.  Now it resets properly.

Note that I'm not doing this for the picture poster.  That already checks
to see if there's a *picture* selected, and you should be allowed to send
a picture without text.


Conflicts:

	src/net/exclaimindustries/geohashdroid/WikiMessageEditor.java

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@596 7ebfa8b6-0479-11de-af5c-836599625e1f