Fixed WikiPictureHandler.
First, I forgot to actually assign the filename when parsing through the
Intent.  Whoops.

Second, WikiPictureHandler now ignores the phoneTime parameter.  Since we
don't actually sign picture uploads, we're only using the time as a
pseudorandomish (kinda sorta) stamp on the filename to avoid confusion.
The way it ran before, the filename could wind up ending with "~~~~~.jpg",
which MediaWiki apparently chokes on.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@601 7ebfa8b6-0479-11de-af5c-836599625e1f