Changed the large main screens to use DateButtons.
As it turns out, the bug I fixed for DatePicker isn't in all devices, as
there's still bug reports coming in about the date not matching what it
says at times.  So I've made large use the same DateButtons that the small
screens use, as I know that shouldn't have the same problems.

To that end, DateButton is also more blinged up, complete with a custom
XML attribute to decide what sort of date should be shown, as there were a
few layout problems in the landscape view.

git-svn-id: https://geohashdroid.googlecode.com/svn/branches/queueservice@606 7ebfa8b6-0479-11de-af5c-836599625e1f