Removed (hopefully) all reference to DatePicker from the main activity.
Since we're not dealing with DatePicker anymore, we no longer need the
catches and special cases in the main activity that tells it whether
we're dealing with a large/medium screen (DatePicker) or a small screen
(DateButton).  So, those calls have been removed to clean up the code a
bit.

git-svn-id: https://geohashdroid.googlecode.com/svn/trunk@608 7ebfa8b6-0479-11de-af5c-836599625e1f