Started the tablet branch.
This branch will hold the Honeycomb interface to Geohash Droid.  I'm not
bothering with the compatibility library (mostly because I still have users
back on 1.5), so anything that happens here is all Honeycomb, all the time.
This interface will be Fragment-based and should look very less ugly under
tablets.

Note that this defaults to API level 11.  So, when compiling anything from
this branch, you MUST use the 3.0 (with Google Maps API) compile options.

The main reason I'm doing this NOW is because I actually have a tablet I
can test the durn thing on.