Made ALL coordinate readouts use periods for decimal delimiters.
After doing a bit of research online, it seems that, even in countries that
use commas as the decimal delimiter, coordinates are ALWAYS given with a
period instead.  If someone can show me definite evidence once way or
another, I'd like to see it just to be sure, but for now, I'm just setting
all coordinate readouts to use en_US DecimalFormatSymbols.