Added the Radar button to the main map screen.
Radar goes in the fifth slot, after Wiki and before Send To Maps.  This
will push Send To Maps into the spillover, but I find this acceptable due
to the fact that tapping the current final destination will bring up the
same prompt anyway.

This is all thanks to the power of RADAR!!!