Changed how the info boxes are organized.
Now, instead of the info boxes just being TextViews with carriage returns,
they're LinearLayouts with numerous TextViews inside them.  That is, each
line is a separate TextView.  This allows me to do more interesting things
with the data, such as changing the text color if you're within a certain
range of the destination point.  That'll come in next.

Also, the graticule map's infobox will be changed soon to match.