Display the current location on the wiki picture upload screens.
Based on a suggestion by Jonas Graudums, the location that will be uploaded
along with wiki-posted pictures (as well as stamped on the picture, if
desired) will be displayed on the screen below the image itself.  This can
help to make sure that you're actually sending an image with the right
coordinates (or, as may be more likely, that the image you're sending
actually HAS coordinates embedded in it).

This hasn't been tested wiki-side yet.  And I still want to do the same
thing with message posting, too.