Created ClosenessActor.
This is the current solution I have for making a Toast popup exactly once
per session when the user is close enough to the point.  It's sort of ugly
and uses SharedPreferences to keep the boolean handy across all Activities
(and looks like it ought to all be static), but for the time being, this is
the basic framework.  I'll add it in to the needed classes soon.