Moved the wiki signature date formatter to DateTools.
The date formatter for wiki signatures is now in DateTools, not in
WikiMessageEditor.  Seems to fit better there.  Hopefully.

Also, changed that same formatter to force it to always use English when
making the date.  This means it matches what WOULD show up on the wiki
itself, as opposed to whatever locale the phone is in.