Added Intent-checking to Send-To-Map mode.
While it's not very likely on any retail phone, I can envision SOME custom
firmware somewhere lacking the Maps app (or any app that can handle the
geo: scheme).  So, just to bulletproof things, we'll check to make sure
we can send the Intent before we offer the user that option.