More fleshing out of StockService.
Ha!  Thought I was dead, did ya?

StockService is now fleshed out with a lot more of what'll make it what it
is.  It'll know to attempt to reconnect if the network goes down while
refreshing, it'll get both stock prices for the day (30W and non-30W) if
need be, and it's just all coming together slowly.