Set StockService's time zone to America/New_York.
Although it's usually the case that any global use of a time zone should
default to UTC (either out of the responsibility involved with global
distribution or just to screw up everyone outside of London equally), in
this case we're dealing with a service (the NYSE) that is tightly bound
to a single time zone (EST) with DST adjustments.  So it's best that we
make all our time-based checks around THAT time zone, rather than shoot
for UTC and hope for the best.