Added convenience methods for getting 9:30am ET.
New idea.  Always retrieve stocks from days we're confident will have
stocks on them.  To that end, the methods here will get "today's" 9:30am
ET and the most likely date that would have a stock value (that is, if
the time is past 9:30am on that date).  That'll get integrated into
things later.