Fixed several issues with StockService.
Turns out I can't pre-make an AlarmManager and NotificationManager in that
manner.  Oops.

Also, unregistering a receiver that was never registered in the first place
throws an exception.  Since that's how I'm handling STOCK_CANCEL_ALARMS,
that call is now try-catched.

Plus, the debug log for the alarm time's initial setup should be a lot more
helpful.