Fixed the network receiver and rearranged STOCK_INIT.
First, I was not aware that destroying a service takes down all its
manually-registered BroadcastReceivers.  That sort of makes the
NetworkReceiver useless (the service is destroyed after registering for
updates) and leaky (the receiver isn't unregistered at onDestroy time).  So
first off, this makes sure we don't stopSelf on StockService when we're in
the middle of waiting for the network.

Also, moved the alarm registration to STOCK_INIT handling.  Because that
seems to make a lot more sense.