Made StockService go to foreground mode when waiting for the network.
As it turns out, the service will most likely be killed if it goes to the
background and stops doing anything, even if it's waiting for a network
connection update.  So, now StockService will remain in the foreground (and
thus less likely to be killed) until the network comes back.  It's still
not DOING anything during that time, but its network receiver will stick
around.