Gave both Notifications Intents to send off.
Turns out the older versions of Notification REQUIRE an Intent to fire off
from Notifications, else it blows up with an IllegalStateException.  So to
that end, the "In progress" Notification starts the main GHD Activity,
while the "Waiting for network" one will, in due time, fire a popup that
will ask the user if it should just give up on waiting for the network and
clear the Notification.