Added strings for canceling "wait for network".
In due time, these strings will be put into a dialog which will pop up
if the user taps the "Waiting for network" notification.  Said dialog
will, as the strings imply, allow the user to tell GHD to give up on
finding a network (clearing the notification) or go to preferences to
shut off prefetch altogether.