Added a dialog for giving up on the network.
If the "Waiting for network" notification is tapped, now the user will get
a dialog asking if GHD should give up on the network for now.

Note that this isn't fully working yet.  I still want it to not call up
whatever the last state of the app was when calling the dialog, for a
start.  Also, the PendingIntent being used seems to only work once, then it
starts throwing exceptions.  Must do more research.