Really really improved how network connectivity is handled.
Oh, so THAT'S how that's supposed to work!

StockService's NetworkReceiver is now registered in the manifest and
controlled via PackageManager calls to enable/disable it.  The upshot of
this is that the service doesn't need to remain up and in the foreground
when we're waiting for the network, meaning we don't need the notification
to be up, meaning DialogHandler goes away, plus the foreground mode lock
isn't needed, meaning ForegroundCompatService dies, plus all the supporting
stuff each needed, not to mention the issues I was having with
DialogHandler in the first place...

Look, it's just a whole lot better, okay?  While I tested all this and it
seems to work the way I wanted it to when I first had this idea, there
might be one or two cases I still need to mess with.