Direct the STOCK_ALARM_RETRY broadcast to the correct receiver.
Okay, THIS time it worked.  Sending broadcasts like that seems more tricky
than it should be, but when I force it to go directly to the appropriate
class, it starts working again.  Which I guess is all right, I can't
imagine anything else wanting to know when StockService is retrying a stock
check.