Added the now-enforced READ_EXTERNAL_STORAGE permission.
Wow.  I really, REALLY need to get to work on the modern remake at some
point.  At any rate, without this permission, anything related to the
picture uploader will fail as it refuses to open images from the SD card.

This will probably be reverted later; I'm going to convert it to just using
the content resolver so this permission probably won't be needed, and it'll
allow for uploads from Picasa, should the need arise.  It shouldn't arise,
though, as that would be sort of redundant in a way, but still.