Added in the service testing Activity.
I may as well toss in ServiceTester for the time being.  I say this because
I think I accidentally pushed a version of the manifest that requires it
and it won't build without it.  And because now you, too, can test the
service.

ServiceTester will go away before release, of course.