Initial commit of AlarmService.
AlarmService replaces the main functionality of the old StockService (and
confusingly uses the NEW StockService to do its job).  This will be what
does the stock pre-fetching at 9:30am every day.

Note that at this point I think the actual app is more than a bit
non-functional in this branch, so don't try to do anything with it.  I'll
get something in ServiceTester to test AlarmService soon.