Initial commit of WikiService.
WikiService is the background service that'll take care of all wiki
connections.  This will replace the "popup and wait until done" nature of
the current wiki activities, making the whole thing cleaner and much
better-suited to things like queuing up wiki posts for later if there's no
network connection.

This is just the base service with stubs for the abstract methods.  More to
follow later, once I've written it.