The Big Bad Android Studio Import That'll Break Everything Commit(tm).
It's a new UI, a new repository host, a new UI, a new Android version, a
new backend, so it follows it's time for a new IDE.  Seems like Google's
going straight to Android Studio, so hey, may as well start learning it on
a personal project, right?  Right.  This commit does that import.

Of course, I'm not quite as familiar with Android Studio/IntelliJ as I am
with Eclipse, so I'm dumping it in a new branch for the time being.  I
probably missed something really super vital required for importing
everything, but hopefully I can get this all sorted out eventually.  I'll
make commits to both here and the main tablet branch for a while until I
know what I'm doing.