Further WikiService work for image uploading.
The start of image processing in WikiService is underway.  I'm largely
pulling things out of WikiPictureEditor to make this happen, along with
some servicey Notifications to go with it.

Also, now that I think about it, EXTRA_PROGRESS shouldn't be a thing, since
the only condition in which we need to track where we were is if we already
uploaded an image (that and the message are the only times we actually
WRITE to the server), and we'll have checks to make sure we didn't already
send the image, just like we had in WikiPictureEditor.