Created resolveWikiExceptionActions.
Since pretty well any problem we get can be boiled down to a wiki problem
(in which case we'll get an error ID) or not (in which case we default to
shrugging and offering generic options), there's now a method that will
fill in the appropriate actions and buttons to any pausing error
notification.  Granted, these actions haven't been made yet, but the basic
framework's there (and the generic error's there, sans icons).

Man, this is going to be HILARIOUS when I finally have enough of it done to
run actual tests...