Lots of WikiService fixes.
Okay, so the bad news is that WikiService had a lot of problems the way I
wrote it, as evidenced when I tested it with the new additions to
ServiceTester.

The good news is that I finally got around to testing it, and so most of
those issues should now be fixed.  Yay!  Please don't use ServiceTester
without changing the wiki URL in WikiUtils, else you WILL be posting to the
real Geohashing wiki, and they won't like that very much.

Note to self: Make sure the image info stamper works right.