Some infobox rendering tweaks.
Turns out "6pt" isn't a guaranteed physical size, at least not as far as
the emulator's concerned.  So, now it's 20px, which I hope clears things up
a bit.

Also, 75 quality for the JPEGs?  We ought to be able to crank that to 85
and still be within size constraints, right?

I may need to completely overhaul the infobox's appearance.  It's not
exactly showing up the way I was hoping it would, especially not when it's
saying Location Unknown.