Expunged PreferenceEditScreen.
It and everything related to it are now gone.

Of course, the Activities I just changed should probably ALSO go away, but
not yet.  Not until I'm sure I've got everything I need out of them.  Which
shouldn't be much.  I'm just making sure things continue to compile right
for the time being.