Put CentralMap where it should be, and made it a map.
Now we've at least got a map.  It doesn't do anything yet, but it IS a map,
and you can't take that from it.

Note, of course, that if you want to compile this yourself, you're going to
need your own API key for Maps v2.  Make it a string resource called
api_map_key_v2, and it should work.