Put the final destination marker on the map.
If CentralMap gets an Info in from StockService now (and it was an Info it
requested), it'll now put an appropriate marker on the map.  And it'll be
labeled properly when the user taps on it, AND the standard GoogleMap
methods will put up navigation and send-to-Maps buttons.  Note that the
send-to-Maps feature appears to not work if the point is in the middle of
the Pacific Ocean, as the globalhash was when I tested this.

Also note that this won't actually get the hashpoints yet without outside
intervention.  You'll have to wire your own test stuff in to get that.