Made CentralMap's StockReceiver only check IDs for user-initiated calls.
I think I've got a plan for how the interface might work.  It'll require a
couple changes to the flags that get sent back and forth to StockService,
and this is the first part of it.  Now, only user-initiated checks will
require a matching ID, and only user-initiated checks will trigger error
messages.