Added initial nearby point drawing.
Nearby points will now be drawn on the map if preferences allow it.

I still need proper text on these, as well as something that should happen
if the user taps the info window (i.e. switch to that graticule).  Also,
the nearby points should vanish if the zoom distance is too far away to be
useful (similar to clustering, I guess).