Made tapping a nearby hashpoint's info window prompt to switch.
This behaves like the old Geohash Droid did, except for the fact that it
requires a second tap (the info window, which pops up after the marker is
tapped).  I'm still undecided on if I want to go back to the old style
where tapping on the marker goes directly to the prompt.