Converted the stock database wipe dialog into a DialogFragment.
Note that I'm not converting the wiki username/password update dialog into
a DialogFragment since that's already an Activity, owing to how it's used.
Also, NearbyGraticuleDialogFragment is its own class file because I get the
distinct feeling CentralMap is going to get way too big and unwieldy as it
is now.