Rearranged the graticule picker a bit.
First, the close button is now in line with the lat/lon fields and the
Globalhash checkbox is next to the Closest button.  This is so that the
close button remains visible when the keyboard is up.  Less confusing.

Also, the container itself is clickable.  This makes it eat clicks, which
in turn means the underlying GoogleMap won't do get them.

The EditTexts are set to flagNoExtractUi (it won't go to a fullscreen
keyboard if you go to landscape mode).  Lat now has actionNext on it (goes
to Lon), and Lon has actionDone (closes the keyboard).

THAT should clear up some of the confusion I was having when I was messing
around with this.