Set CentralMap to a windowSoftInputMode of stateHidden.
There's a minor issue with the graticule picker, in that if the keyboard is
open during a rotation (or whatever else would trigger an instance save),
it comes back up but corrupts some of the screen in portrait mode until it
goes away.  So now the keyboard will just not stay open on a rotation.

Though I may change it so the graticule picker appears on the TOP, not the
BOTTOM.  That may also fix it, but I like it on the bottom.