Did a lot of work towards the CentralMap graticule picker interface.
There's still a ways to go, given CentralMap has to work in two separate
modes (Select-A-Graticule and Expedition), but this is a start.  I still
need to keep track of what to do with Info objects as they come in and when
the "current" one becomes an actual expedition.  Also, possible zooming
issues regarding the graticule picker being on-screen in things like
landscape mode.  That's a killer.