Made CentralMapMode.
So, having built up CentralMap around switching between expedition and
Select-A-Graticule modes, I realized the way I did it was incredibly
wrong, in that a bunch of methods in CentralMap had to switch on a boolean
as to whether or not they were to be handled as one mode or another.  That
was starting to bloat CentralMap badly and make a lot of confusing,
terrible code.

So now I'm in the process of breaking those modes out into CentralMapMode
objects that live outside of CentralMap.  I don't have the concrete
implementations finished quite yet (and CentralMap doesn't call them), but
this is the basic structure.  More to come.