Fixed initial zooms in ExpeditionMode.
The way I phrased the preference, initial zoom should always happen when
entering expedition mode.  The preference should only be checked when the
map is resumed for whatever reason while already in that mode.

Though to be honest, I might want to make sure that the preference-
controlled situation actually restores the old camera when leaving
SelectAGraticuleMode, not necessarily always go to initial zoom.  Hm.