Separated bundle saving from cleanUp in CentralMapMode.
Apparently onSaveInstanceState can be called in an Activity even when it's
just onStopping, not when the Activity is definitely going to be killed.  I
was not aware of this.  Because cleanUp in CentralMapMode was ALWAYS the
way the Bundle was saved, and said Bundle was being saved in
onSaveInstanceState, this meant the mode was being wiped clean but never
replaced later by onCreate, leaving things in a goofy state.

So now CentralMapMode has an onSaveInstanceState method of its own.  That's
where the Bundle-writing happens, and cleanUp only does cleanup tasks, and
is only called when the mode is definitely going down.