Initial commit of GHDDatePickerDialogFragment.
As I said in the JavaDoc, this wouldn't be necessary if not for the desire
to stick a "Today" button in the dialog.  That makes it so I can't just use
DatePickerDialog.  There's still a few hitches here and there, which I
ought to be able to clean up soon.

So now both modes will react to new dates, which is good.  Now I need to
settle on a startup routine.  Something that'll ensure there shouldn't be a
situation where there's no hashpoint set.  Hm...

Also, added more @Nullable and @NotNulls around.  Those seem useful.  For
now.