Moved nearby point fetching into StockService.
I know I kept telling myself I wanted to keep StockService as simple as
possible, but it turns out I'll need the entire "fetch the infos for
nearby graticules" logic for more than just one spot.  So now StockService
can be told to fetch the nearby points and return them as an array of
Parcelables along with the main result.