Made Info.measureClosest throw an exception for bad input.
Leaving Info as @Nullable was starting to get confusing in places where I
needed to keep checking the output just to make sure it came back not-null.
So, now it's explicitly @NonNull, and to do so, it'll throw an exception if
the single Info is null AND the array is either null or empty.