Added a better startup routine for CentralMap.
After much debate with myself as to what, exactly, CentralMap SHOULD do on
initial startup, I came to the conclusion that, unless it has a saved
instance state (which would make it not be an initial startup, I guess),
CentralMap will tell ExpeditionMode to get the closest Info for the day's
hash (assuming it's been posted yet) and start with that.  I'm pretty sure
that's by far the most common use of the app.

Now, this will have to be revisited once I get something like
TracklogService running, as that may mean an expedition is "active" when
the app is closed, but for now, this'll do.

I'm also amazed it seemed to work perfectly the first time I put it on my
tablet, which makes me think it's subtly broken very badly and I'll learn
that the hard way later.