Fixed Select-A-Graticule not rebuilding properly from screen rotations.
Select-A-Graticule wasn't reassigning itself as a listener to the fragment
if said fragment wasn't being explicitly created by SelectAGraticuleMode.
That is, if the fragment was part of the saved state (i.e. if the display
rotates and has to be taken down), the fragment would've been rebuilt by
the Activity, meaning mFrag.setListener would never be called.  Now that
should all be settled.