Fixed potential startup problem.
I don't honestly know if this is possible to trigger in production, but I
ran into a few problems with the empty start when sending the APK to my
device when the screen was off.  Everything started up, but
ExpeditionMode's empty startup (where it automatically finds the nearest
Graticule for today) wasn't firing off.

To fix that, CentralMap will now only ask the current mode for its instance
state bundle if the mode's init has completed.  If it hasn't, CentralMap
will just retain whatever was the last bundle it knew about.