Redid the final destination graphics.
There.  Now I've got a clean SVG of the final destination flag.  As such, I
also have a new set of graphics for it, now featuring xhdpi and xxhdpi
versions.

The disabled version (nearby points) has been completely redone.  Now, in
addition to being a translucent image, the flag isn't waving in the
imaginary wind that just happens to be around, and is thus crumpled up
against the pole.

I didn't touch the smaller version yet; I've got other ideas for that.  And
as soon as I decide on a good place to dump the original SVGs, I'll put
those in the repo as well.  Man, it'd be nice if Android got native SVG
drawable support...