Initial commit of the new InfoBox.
There!  Now it shows up and does something in ExpeditionMode.  I realize
now that I left out the accuracy warning and greening the distance if the
user is close enough, but that'll come next.  What'll also come next is
a way to remove the infobox (i.e. animating it away) and making sure it
doesn't show up if the user didn't ask for it.

As a side note, don't expect the Jumbo Mode Infobox to return.  Also, note
that isn't the same as the detailed info screen, which WILL return, though
maybe in a Fragment instead of always a new Activity.